requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-reset-migrations==0.4.0
django-widget-tweaks==1.4.12
idna==3.4
oauthlib==3.2.1
Pillow==9.2.0
pycparser==2.21
PyJWT==2.5.0
python-decouple==3.6
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
types-cryptography==3.3.23
tzdata==2022.2
urllib3==1.26.12
